fix: Geminiレビュー指摘事項の修正（2点）
1. DTO定義をプレゼンテーション層に分離
   - TransactionDto, SubcategoryAggregationResponseDto, TrendDataResponseDto, CategoryAggregationResponseDtoをget-category-aggregation.dto.tsに移動
   - Onion Architectureの原則に従い、関心事を分離

2. executeメソッドのパフォーマンス改善
   - ループの前にカテゴリタイプ別に取引をグルーピング（Map<CategoryType, Transaction[]>）
   - allTotalAmountをループ外で一度だけ計算
   - 各集計メソッドにフィルタリング済みの配列を渡すことで、不要な反復処理を削減
   - aggregateByCategoryTypeメソッドのシグネチャを変更（allTotalAmountを引数で受け取る）

3. ルールへの取り込み
   - 詳細設計ガイドラインにループ処理全体の最適化の観点を追加

Issue: #47
