+++ apps/backend/src/modules/transaction/presentation/dto/get-category-aggregation.dto.ts
@@ -16,3 +16,55 @@ export class GetCategoryAggregationQueryDto {
   @IsEnum(CategoryType)
   categoryType?: CategoryType;
 }
+
+/**
+ * TransactionDto
+ * プレゼンテーション層用のDTO
+ */
+export interface TransactionDto {
+  id: string;
+  date: string; // ISO8601形式
+  amount: number;
+  categoryType: string; // CategoryTypeの文字列値
+  categoryId: string;
+  institutionId: string;
+  accountId: string;
+  description: string;
+}
+
+/**
+ * SubcategoryAggregationResponseDto
+ */
+export interface SubcategoryAggregationResponseDto {
+  categoryId: string;
+  categoryName: string;
+  amount: number;
+  count: number;
+  percentage: number;
+  topTransactions: TransactionDto[];
+}
+
+/**
+ * TrendDataResponseDto
+ */
+export interface TrendDataResponseDto {
+  monthly: Array<{
+    month: string; // YYYY-MM
+    amount: number;
+    count: number;
+  }>;
+}
+
+/**
+ * CategoryAggregationResponseDto
+ */
+export interface CategoryAggregationResponseDto {
+  categoryType: CategoryType;
+  startDate: string; // YYYY-MM-DD
+  endDate: string; // YYYY-MM-DD
+  totalAmount: number;
+  transactionCount: number;
+  subcategories: SubcategoryAggregationResponseDto[];
+  percentage: number;
+  trend: TrendDataResponseDto;
+}
