+++ docs/development/detailed-design-guideline.md
@@ -497,6 +497,10 @@ docs/detailed-design/
   - ループ内で配列の`filter`を実行すると計算量がO(取引数 × サブカテゴリ数)になる
   - 事前にMapでグルーピングすることで、計算量をO(取引数)に改善できる
   - 例: サブカテゴリごとの取引を取得する際、ループ前に`Map<subcategoryId, Transaction[]>`を作成する
+  - **ループ処理全体の最適化**: カテゴリタイプごとのループ内で、同じ配列に対して複数回フィルタリングを実行するのは非効率
+  - ループの前に、全取引をカテゴリタイプ別に`Map<CategoryType, Transaction[]>`にグルーピングし、ループ内ではフィルタリング済みの配列を使用する
+  - 全体の合計金額（`allTotalAmount`）もループ外で一度だけ計算し、引数として渡すことで、不要な反復処理を削減する
+  - 例: `execute`メソッドで、ループ前に`transactionsByCategoryType`を作成し、各集計メソッドにフィルタリング済みの配列を渡す
 
 - [ ] **Enum値の動的取得**
   - ハードコードされたEnum値のリストは、Enumが変更された際に追従漏れが発生する可能性がある
